refactoring of cards fetching process
