+++ webapp/src/app.d.ts
@@ -26,34 +26,16 @@ declare global {
 		hex_color: string;
 	}
 
-	// TODO: Add category_id to the Card table in supabase
 	interface Card {
 		id: string;
         name: string;
         year: number;
 		creator: string;
-		picture_path: string;
+		picture_url: string;
 		category_id: string;
 		game_id: string;
-		player_id: string;
-	};
-
-	interface Player {
-		id: string;
-		name: string;
-		avatar_path: string;
-		is_ready: boolean;
-		cards_count: number;
-		game_id: string;
-	}
-
-	interface Game {
-		id: string;
-		status: string;
-		max_card_count: number;
-		difficulty: number;
-		category_id: string;
-		whose_turn_id: string;
+		player_id?: string;
+		in_deck?: boolean;
 	}
 
 	interface Player {
